feat(kumactl): get rid of dead code
