+++ pkg/transparentproxy/kubernetes/kubernetes_test.go
@@ -23,7 +23,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/kumahq/kuma/pkg/plugins/runtime/k8s/metadata"
-	"github.com/kumahq/kuma/pkg/transparentproxy/config"
 	"github.com/kumahq/kuma/pkg/transparentproxy/kubernetes"
 )
 
@@ -222,151 +221,6 @@ var _ = Describe("kubernetes", func() {
 				"--ebpf-instance-ip", "$(FOO_BAR)",
 				"--ebpf-bpffs-path", "/baz/bar/foo",
 				"--ebpf-programs-source-path", "/foo",
-			},
-		}),
-	)
-
-	type testCaseTransparentProxyConfig struct {
-		pod      *kube_core.Pod
-		tpConfig *config.TransparentProxyConfig
-	}
-
-	DescribeTable("should generate transparent proxy config", func(given testCaseTransparentProxyConfig) {
-		podRedirect, err := kubernetes.NewPodRedirectForPod(given.pod)
-		Expect(err).ToNot(HaveOccurred())
-
-		tpConfig := podRedirect.AsTransparentProxyConfig()
-		Expect(tpConfig).To(Equal(given.tpConfig))
-	},
-		Entry("should generate", testCaseTransparentProxyConfig{
-			pod: &kube_core.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{
-						metadata.KumaBuiltinDNS:                                 metadata.AnnotationEnabled,
-						metadata.KumaBuiltinDNSPort:                             "25053",
-						metadata.KumaTrafficExcludeOutboundPorts:                "11000",
-						metadata.KumaTransparentProxyingOutboundPortAnnotation:  "25100",
-						metadata.KumaTrafficExcludeInboundPorts:                 "12000",
-						metadata.KumaTransparentProxyingInboundPortAnnotation:   "25204",
-						metadata.KumaTransparentProxyingInboundPortAnnotationV6: "25206",
-						metadata.KumaSidecarUID:                                 "12345",
-					},
-				},
-			},
-			tpConfig: &config.TransparentProxyConfig{
-				DryRun:                 false,
-				Verbose:                true,
-				RedirectPortOutBound:   "25100",
-				RedirectInBound:        true,
-				RedirectPortInBound:    "25204",
-				RedirectPortInBoundV6:  "25206",
-				ExcludeInboundPorts:    "12000",
-				ExcludeOutboundPorts:   "11000",
-				UID:                    "12345",
-				GID:                    "12345",
-				RedirectDNS:            true,
-				RedirectAllDNSTraffic:  false,
-				AgentDNSListenerPort:   "25053",
-				DNSUpstreamTargetChain: "",
-			},
-		}),
-		Entry("should generate experimental engine", testCaseTransparentProxyConfig{
-			pod: &kube_core.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{
-						metadata.KumaTrafficExcludeOutboundPorts:                "11000",
-						metadata.KumaTrafficExcludeOutboundUDPPortsForUIDs:      "11001:1;11002:2",
-						metadata.KumaTrafficExcludeOutboundTCPPortsForUIDs:      "11003:3",
-						metadata.KumaTransparentProxyingOutboundPortAnnotation:  "25100",
-						metadata.KumaTrafficExcludeInboundPorts:                 "12000",
-						metadata.KumaTransparentProxyingInboundPortAnnotation:   "25204",
-						metadata.KumaTransparentProxyingInboundPortAnnotationV6: "25206",
-						metadata.KumaSidecarUID:                                 "12345",
-						metadata.KumaTransparentProxyingExperimentalEngine:      metadata.AnnotationEnabled,
-					},
-				},
-			},
-			tpConfig: &config.TransparentProxyConfig{
-				DryRun:                         false,
-				Verbose:                        true,
-				RedirectPortOutBound:           "25100",
-				RedirectInBound:                true,
-				RedirectPortInBound:            "25204",
-				RedirectPortInBoundV6:          "25206",
-				ExcludeInboundPorts:            "12000",
-				ExcludeOutboundPorts:           "11000",
-				UID:                            "12345",
-				GID:                            "12345",
-				RedirectDNS:                    false,
-				RedirectAllDNSTraffic:          false,
-				AgentDNSListenerPort:           "0",
-				DNSUpstreamTargetChain:         "",
-				ExperimentalEngine:             true,
-				ExcludeOutboundUDPPortsForUIDs: []string{"11001:1", "11002:2"},
-				ExcludeOutboundTCPPortsForUIDs: []string{"11003:3"},
-			},
-		}),
-		Entry("should generate no builtin DNS", testCaseTransparentProxyConfig{
-			pod: &kube_core.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{
-						metadata.KumaTrafficExcludeOutboundPorts:                "11000",
-						metadata.KumaTransparentProxyingOutboundPortAnnotation:  "25100",
-						metadata.KumaTrafficExcludeInboundPorts:                 "12000",
-						metadata.KumaTransparentProxyingInboundPortAnnotation:   "25204",
-						metadata.KumaTransparentProxyingInboundPortAnnotationV6: "25206",
-						metadata.KumaSidecarUID:                                 "12345",
-					},
-				},
-			},
-			tpConfig: &config.TransparentProxyConfig{
-				DryRun:                 false,
-				Verbose:                true,
-				RedirectPortOutBound:   "25100",
-				RedirectInBound:        true,
-				RedirectPortInBound:    "25204",
-				RedirectPortInBoundV6:  "25206",
-				ExcludeInboundPorts:    "12000",
-				ExcludeOutboundPorts:   "11000",
-				UID:                    "12345",
-				GID:                    "12345",
-				RedirectDNS:            false,
-				RedirectAllDNSTraffic:  false,
-				AgentDNSListenerPort:   "0",
-				DNSUpstreamTargetChain: "",
-			},
-		}),
-		Entry("should generate for Gateway", testCaseTransparentProxyConfig{
-			pod: &kube_core.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{
-						metadata.KumaBuiltinDNS:                                 metadata.AnnotationEnabled,
-						metadata.KumaBuiltinDNSPort:                             "25053",
-						metadata.KumaTrafficExcludeOutboundPorts:                "11000",
-						metadata.KumaTransparentProxyingOutboundPortAnnotation:  "25100",
-						metadata.KumaGatewayAnnotation:                          metadata.AnnotationEnabled,
-						metadata.KumaTrafficExcludeInboundPorts:                 "12000",
-						metadata.KumaTransparentProxyingInboundPortAnnotation:   "25204",
-						metadata.KumaTransparentProxyingInboundPortAnnotationV6: "25206",
-						metadata.KumaSidecarUID:                                 "12345",
-					},
-				},
-			},
-			tpConfig: &config.TransparentProxyConfig{
-				DryRun:                 false,
-				Verbose:                true,
-				RedirectPortOutBound:   "25100",
-				RedirectInBound:        false,
-				RedirectPortInBound:    "25204",
-				RedirectPortInBoundV6:  "25206",
-				ExcludeInboundPorts:    "12000",
-				ExcludeOutboundPorts:   "11000",
-				UID:                    "12345",
-				GID:                    "12345",
-				RedirectDNS:            true,
-				RedirectAllDNSTraffic:  false,
-				AgentDNSListenerPort:   "25053",
-				DNSUpstreamTargetChain: "",
 			},
 		}),
 	)
